requirements.txt
ipykernel
jupyter
openai
pandas
pymupdf
faiss-cpu
tqdm
tenacity
